Enable Operation._add_control_inputs() with the C API and related improvements
This change:
 - Implements the C API logic for Operation._add_control_inputs()
 - Adds type-checking to Operation._add_control_input()
 - Makes Graph::AddControlEdge() optionally update the node def

The AddControlEdge() changes may have a performance impact if anything
is sensitive to AddControlEdge(), but nothing is to my knowledge. I'm
not sure what benchmarks would confirm this.

This changes all users of AddControlEdge() to update the node def,
except for a few cases that explicitly require not doing this
(e.g. importing a graph def). This should make the graph state more
consistent in the other use cases (tests don't fail at least).